// frontend/src/app/guest-room/types/webrtc-types.ts
export interface MediaDevice {
  deviceId: string;
  kind: 'audioinput' | 'audiooutput' | 'videoinput';
  label: string;
  groupId: string;
}